Use of preset files to allow backups
The logic will be as follows. The script will read in configuration
information from preset files. Preset files are INI files containing the
name of the preset, the source and the destination. A simple example of
such a preset file would be:

[home]
source = /home/user/
destination = /mnt/backup/

The script will then need to copy over the files in the source into the
destination under a folder named after today's date (e.g., 2014-08-10/)
and create a link to it marking it as the latest backup. Subsequent
backups will be added alongside and each time, the link will point to
the latest backup.

To do this, first, the input arguments must be parsed. Next, each valid
preset file is processed and each valid preset is archived.