Preserve extended attributes of files
The --archive option stands for -rlptgoD where the individual options
stand for the following
--recursive
--links
--perms
--times
--group
--owner
--devices
--specials

This set of options preserves the permissions, links, group, owner and
device and special files. However, it explicitly excludes the following
--hard-links
--acls
--xattrs

While avoiding hard links is a sound idea, avoiding extended attributes
stored by the file system is not. In KDE, baloo, the semantic data
subsystem stores all the ratings, tags, comments, etc. in the extended
attributes. I want to include those when I make a backup. So, this
commit includes the --xattrs option explicitly.